movies repository implemented
