+++ app/src/main/java/com/example/movielister/api/MoviesAPI.kt
@@ -1,8 +1,10 @@
 package com.example.movielister.api
 
-import com.example.movielister.BuildConfig.*
+import com.example.movielister.BuildConfig.API_KEY
+import com.example.movielister.BuildConfig.API_STR
 import com.example.movielister.model.MovieModel
 import com.example.movielister.model.MovieResponseModel
+import com.example.movielister.network.WebService
 import retrofit2.http.GET
 import retrofit2.http.Path
 
@@ -15,4 +17,8 @@ interface MoviesAPI {
 
     @GET("$MOVIE{id}$API_STR$API_KEY")
     suspend fun fetchMovie(@Path("id") movieId: Int): MovieModel?
+}
+
+object MoviesNetwork {
+    val movieService: MoviesAPI = WebService.createRetrofit().create(MoviesAPI::class.java)
 }
