+++ app/src/main/java/com/example/movielister/movies/MoviesViewModel.kt
@@ -2,28 +2,21 @@ package com.example.movielister.movies
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.example.movielister.model.MovieModel
-import com.example.movielister.network.MoviesNetwork
-import kotlinx.coroutines.flow.MutableSharedFlow
+import com.example.movielister.repository.MoviesRepository
 import kotlinx.coroutines.flow.asSharedFlow
 import kotlinx.coroutines.launch
 
 class MoviesViewModel() : ViewModel() {
-    private val moviesService = MoviesNetwork.createMoviesAPI()
+    private val moviesRepository = MoviesRepository()
 
-    private val _popularMoviesList = MutableSharedFlow<List<MovieModel>>()
-    val popularMoviesList = _popularMoviesList.asSharedFlow()
+    val popularMoviesList = moviesRepository._popularMoviesList.asSharedFlow()
 
-    private val _currentMovie = MutableSharedFlow<MovieModel>()
-    val currentMovie = _currentMovie.asSharedFlow()
+    val currentMovie = moviesRepository._currentMovie.asSharedFlow()
 
     fun fetchPopularMovies() {
         viewModelScope.launch {
             runCatching {
-                moviesService.fetchPopularMovies()?.let {
-                    val popularMovies = it.results
-                    _popularMoviesList.emit(popularMovies)
-                }
+                moviesRepository.fetchPopularMovies()
             }.onFailure {
                 it.printStackTrace()
             }
@@ -33,10 +26,7 @@ class MoviesViewModel() : ViewModel() {
     fun fetchMovie(movieID: Int) {
         viewModelScope.launch {
             runCatching {
-                moviesService.fetchMovie(movieID)?.let {
-                    val movie = it
-                    _currentMovie.emit(movie)
-                }
+                moviesRepository.fetchMovie(movieID)
             }.onFailure {
                 it.printStackTrace()
             }
